+++ R/H2cal.R
@@ -104,19 +104,20 @@
 #' dt <- potato
 #'
 #' hr <- H2cal(data = dt
-#'             , trait = "tubdw"
+#'             , trait = "stemdw"
 #'             , gen.name = "geno"
 #'             , rep.n = 5
 #'             , fixed.model = "0 + (1|bloque) + geno"
 #'             , random.model = "1 + (1|bloque) + (1|geno)"
 #'             , emmeans = TRUE
-#'             , plot_diag = TRUE
+#'             , plot_diag = FALSE
 #'             , outliers.rm = TRUE
 #'             )
 #'
 #'  hr$tabsmr
 #'  hr$blues
 #'  hr$blups
+#'  hr$outliers
 #'  
 
 H2cal <- function(data
@@ -137,28 +138,24 @@ H2cal <- function(data
                   , trial = NULL
                   ){
   
+  
 
-# test --------------------------------------------------------------------
-
-  if(FALSE) {
-   
-    data <- dt
+# -------------------------------------------------------------------------
+  
+  if (FALSE) {
     
-    trait = "tubdw"
+    data = dt
+    trait = "stemdw"
     gen.name = "geno"
     rep.n = 5
-    env.n = 1
-    year.n = 1
-    env.name = NULL
-    year.name = NULL
-    fixed.model = "1 + (1|bloque) + geno"
+    fixed.model = "0 + (1|bloque) + geno"
     random.model = "1 + (1|bloque) + (1|geno)"
-    summary = TRUE
-    emmeans = FALSE
-    weights = NULL
+    emmeans = TRUE
     plot_diag = TRUE
-    outliers.rm = TRUE
-    trial = NULL
+    outliers.rm = F
+    
+    weights = NULL
+    summary = TRUE
     
   }
   
@@ -215,27 +212,32 @@ H2cal <- function(data
   }
 
 # Plot models -------------------------------------------------------------
-
+  
+  diag.fix <- inti::plot_diag(g.fix, title = paste("Fixed model:", trait))
+  diag.ran <- inti::plot_diag(g.ran, title = paste("Random model:", trait))
+  
+  diag.plot <- list(fixed = diag.fix, random = diag.ran)
+  
   if (plot_diag == TRUE) {
-
-    prp <- par(no.readonly = TRUE)
-    on.exit(par(prp))   
     
-    par(mfrow=c(2,4))
-    hist(resid(g.fix), main = trait)
-    qqnorm(resid(g.fix), main = trait); qqline(resid(g.fix))
-    plot(fitted(g.fix), resid(g.fix, type = "pearson"), main = trait); abline(h=0)
-    plot(resid(g.fix), main = trait)
-    hist(resid(g.ran), main = trait)
-    qqnorm(resid(g.ran), main = trait); qqline(resid(g.ran))
-    plot(fitted(g.ran), resid(g.ran, type = "pearson"), main = trait); abline(h=0)
-    plot(resid(g.ran), main = trait)
-    par(mfrow=c(1,1))
-
-  }
+    grid::grid.newpage()
+    grid::pushViewport(grid::viewport(layout = grid::grid.layout(2, 4)))
+    
+    vplayout <- function(x, y) 
+      grid::viewport(layout.pos.row = x, layout.pos.col = y)
+    print(diag.fix$histogram, vp = vplayout(1, 1))
+    print(diag.fix$qqplot, vp = vplayout(1, 2))
+    print(diag.fix$residual, vp = vplayout(1, 3))
+    print(diag.fix$homoscedasticity, vp = vplayout(1, 4))
+    print(diag.ran$histogram, vp = vplayout(2, 1))
+    print(diag.ran$qqplot, vp = vplayout(2, 2))
+    print(diag.ran$residual, vp = vplayout(2, 3))
+    print(diag.ran$homoscedasticity, vp = vplayout(2, 4))
+    
+    #> https://data.library.virginia.edu/diagnostic-plots/
+    } 
   
-  
-# handle model estimates --------------------------------------------------
+# Model estimates --------------------------------------------------
   
 # number of genotypes
 
@@ -396,7 +398,6 @@ H2cal <- function(data
         dplyr::rename(!!trait := .) %>% 
         mutate(across({{trait}}, as.numeric)) %>% 
         mutate(smith.w = diag(solve(vcov(g.fix)))) %>% 
-        #>
         dplyr::filter(grepl({{gen.name}}, .data[[gen.name]])) %>% 
         mutate(across({{gen.name}}, ~stringr::str_replace(., gen.name, "")))
       
@@ -405,9 +406,7 @@ H2cal <- function(data
         base::as.matrix(.) %>%
         diag(.) %>%
         enframe() %>%
-        #>
         dplyr::filter(grepl({{gen.name}}, .data$name)) %>% 
-        ##>
         select(!.data$name) %>% 
         deframe() %>% 
         mean()
@@ -475,10 +474,13 @@ H2cal <- function(data
     mutate(H2.p = V.g/(V.g + vdBLUEs/2)) %>%
     mutate(vdBLUPs = vdBLUP.avg) %>% 
     mutate(H2.c = 1 - (vdBLUPs/2/V.g)) %>% 
-    select(!matches("BLUE|BLUP")) %>% 
-    purrr::discard(~all(. == 0 | is.nan(.))) %>% 
+    select(!matches("BLUE|BLUP")) %>%
+    purrr::discard(~all(is.nan(.))) %>%
+    {if (env.n == 1) dplyr::select(.data = ., !V.gxl) else .} %>% 
+    {if (year.n == 1) dplyr::select(.data = ., !V.gxy) else .} %>%
+    {if (env.n && year.n == 1) dplyr::select(.data = ., !V.gxlxy) else .} %>%
     {if (!is.null(trial)) dplyr::mutate(.data = ., trial = trial) else .} %>% 
-    {if (!is.null(trial)) select(.data = ., trial, everything()) else .}
+    {if (!is.null(trial)) dplyr::select(.data = ., trial, everything()) else .}
     
 
 # result ------------------------------------------------------------------
@@ -489,6 +491,7 @@ H2cal <- function(data
     , blues = BLUEs
     , model = g.ran
     , outliers = outliers
+    , diagplot = diag.plot
     )
 }