+++ inst/tarpuy/server.R
@@ -44,6 +44,8 @@ shinyServer(function(input, output, session) {
 
 # auth --------------------------------------------------------------------
 
+  if (file.exists("www/analytics.r")) {source("www/analytics.r", local = T)}
+  
   gar_shiny_auth(session)
 
 # longin vs local ---------------------------------------------------------
@@ -92,9 +94,6 @@ shinyServer(function(input, output, session) {
     as_sheets_id( fieldbook_url() )
 
   })
-  
-
-if (file.exists("www/analytics.r")) {source("www/analytics.r", local = T)}
   
 # generate sheet url ------------------------------------------------------