+++ inst/yupana/ui.R
@@ -16,6 +16,8 @@
 source("msgs.R")
 suppressPackageStartupMessages({source("pkgs.R")})
 
+# -------------------------------------------------------------------------
+
 options("googleAuthR.scopes.selected" = c("https://www.googleapis.com/auth/spreadsheets"
                                           , "https://www.googleapis.com/auth/userinfo.email"
                                           ))